meta: Remove `issues.md`. Closes #1
The whole reason for the `issues.md` was to track some issues that
needed to be fixed since upstream was not maintained very well. However,
there is must more activity on that front so there is little reason to
have this file. Eventually, my custom forms should be replaced with the
official upstream (there is likely little to no difference there
anyways)